Changed BDTime to uint64_t. Preparing to change user units to ns
This probably isn't strictly necessary for NEF stuff, but it enables some
experiments with the output flakiness that I want to do. More generally,
it lets you use the full dynamic range of the FPGA time unit

+++ src/bddriver/common/DriverTypes.h
@@ -12,7 +12,7 @@ namespace bddriver {
 ////////////////////////////////////////
 // decoder/encoder
 
-typedef unsigned int BDTime;
+typedef uint64_t BDTime;
 
 // decoder
 struct DecOutput {
